// core/account/src/types/account/GetInfoRes.ts
// Original file: ../proto/account.proto


export interface GetInfoRes {
  'role'?: (string);
  'givenName'?: (string);
  'surname'?: (string);
  'email'?: (string);
  'token'?: (string);
}

export interface GetInfoRes__Output {
  'role'?: (string);
  'givenName'?: (string);
  'surname'?: (string);
  'email'?: (string);
  'token'?: (string);
}